Fucking binding re-direct problems are back! I found this Google group stuff: https://groups.google.com/forum/#!topic/particularsoftware/VKSgQ9lQtog
that deals explicilty with NHibernate causing the issue. None of the GitHub links work whcih is fucking annoying, so I can't see the code that the person who is helping out if referring to

I tried to update the NHibernate binding from:
newVersion="4.1.1.4000"

to:
newVersion="4.1.2.4000"

and that didn't do shit. Going to try some other stuff. This is annoying